like with wav, read all ogg data and use bytes reader
Using a vorbis stream as a reader would cause it to decode
the ogg data more than once. It's more efficient to use a
bytes reader that would produce a decoded bytes stream.

This will consume more memory per ogg resource,
but the performance would increase.

It's especially noticeable in wasm builds (even more so on
mobile devices). Without this, music is played with noticeable
stutters and artifacts.