add AudioInfo.StreamDecorator feature
This is a breaking change: OGG will no longer be wrapped
in an InfiniteLoop by default. Use StreamDecorator=LoopOGG
to get the old behavior.

As for WAVs, it's now possible to disable the in-memory
approach this library has (use NopDecorator).
This can be handy for games with tons of WAVs (hundreds or thousands).
It's still recommended to keep all SFX in-memory for smaller
games and at least keep the most frequent sounds for large games.

This new StreamDecorator feature will be essential to the
upkoming custom audio loaders feature.
It also makes it possible to use custom decorators like
your own infinite loop implementation, etc.